+++ src/domain/action.go
@@ -1,0 +1,17 @@
+package domain
+
+import (
+	"github.com/google/uuid"
+	"time"
+)
+
+
+type ActionInstance struct {
+	ID                 uuid.UUID
+	WorkflowInstanceId uint64
+	Name      string
+	Facts     Facts
+	CreatedAt *time.Time
+	UpdatedAt          *time.Time
+	FinishedAt         *time.Time
+}
