+++ src/domain/build_info.go
@@ -1,0 +1,6 @@
+package domain
+
+var BuildInfo struct {
+	Version string
+	Commit  string
+}
