+++ src/domain/fact.go
@@ -1,0 +1,3 @@
+package domain
+
+type Facts map[string]interface{}
