+++ src/main.go
@@ -8,21 +8,14 @@ import (
 	"github.com/jackc/pgx/v4/pgxpool"
 )
 
-var BuildInfo struct {
-	Version string
-	Commit  string
-}
-
 func Init() (db *pgxpool.Pool, nomadClient *nomad.Client, err error) {
 	db, err = config.DBConnection(os.Getenv("DATABASE_URL"))
 	if err != nil {
 		return
 	}
-	config := nomad.DefaultConfig()
-	nomadClient, err = nomad.NewClient(config)
+	nomadClient, err = config.NomadClient()
 	if err != nil {
 		return
 	}
-
 	return
 }
